requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase
pydantic
